fix what wasnt broken
tile daemon now runs pending
tiling reqs on a scheduler like
a chump. theres gotta be a better way,
but this seems to get the job done.

original worry over missing half
rye laps was forgot to look in laps_edge
there they are.

would like to add an acceleration
gauge to the trip detector. it recorded
a nap on the drive past target stadium.

Signed-off-by: ia <isaac.ardis@gmail.com>